+++ gorgon/toolkit/sse/__init__.py
@@ -1,2 +1,2 @@
-from .sse_hunter import pySSEHunter
-from .sse_helix_correspondence import SSEHelixCorrespondence
+from .hunter import pySSEHunter
+from .helix_correspondence import SSEHelixCorrespondence
